[GH-30] - better conversion checks
