--- src/NSubstitute.Analyzers.Shared/DiagnosticAnalyzers/IndexerInfo.cs
@@ -1,15 +1,0 @@
-namespace NSubstitute.Analyzers.Shared.DiagnosticAnalyzers
-{
-    internal struct IndexerInfo
-    {
-        public bool VerifyIndexerCast { get; }
-
-        public bool VerifyAssignment { get; }
-
-        public IndexerInfo(bool verifyIndexerCast, bool verifyAssignment)
-        {
-            VerifyIndexerCast = verifyIndexerCast;
-            VerifyAssignment = verifyAssignment;
-        }
-    }
-}
